Feature/document event

+++ src/app/Controllers/ApplicationController.php
@@ -28,14 +28,30 @@ class ApplicationController extends BaseController
 
     public function index()
     {
-        // SELECT * FROM program;
-        $query = $this->db->query('SELECT program.program_num, program.name, application.status, application.app_num
-        FROM program JOIN application ON program.program_num = application.program_num;');
+        $user = sessionUser();
+        if(!$user) return $this->response->redirect(site_url('/login'));
+
+        $query;
+        $student_name;
+
+        if($user->User_Type == 'student'){
+            // SELECT * FROM program;
+            $sql = <<<SQL
+                SELECT * FROM application_filter WHERE application_filter.UIN = $user->UIN; 
+            SQL;
+            $query = $this->db->query($sql);
+        }else{
+            $sql = <<<SQL
+                SELECT * FROM application_filter JOIN user ON application_filter.UIN = user.UIN
+            SQL;
+            $query = $this->db->query($sql);
+        }
 
         $data = [
             'page_title' => 'View Applications | TAMU CyberSec Center',
             // use getResultArray() on queries that return multiple rows
-            'applications' => $query->getResultArray()
+            'applications' => $query->getResultArray(),
+            'user' => $user
         ];
 
         return view('application/index', $data);
@@ -43,19 +59,41 @@ class ApplicationController extends BaseController
     
     public function create() {
         // select only programs that the user hasn't applied for already
-        // TODO: use session for UIN = 
-        $query = $this->db->query('SELECT * FROM program WHERE program_num NOT IN (SELECT program_num FROM application WHERE UIN = 230006744);');
+        $user = sessionUser();
+        if(!$user) return $this->response->redirect(site_url('/login'));
 
-        $data = [
-            'page_title' => 'Applications | TAMU CyberSec Center',
-            'programs' => $query->getResultArray()
-        ];
+        $query;
 
-        return view('application/create', $data);
+        if($user->hasPermission('student')){
+            $sql = <<<SQL
+                SELECT * FROM program WHERE program_num NOT IN
+                (SELECT program_num FROM application WHERE UIN = $user->UIN);
+            SQL;
+            $query = $this->db->query($sql);
+
+            $array = $query->getResultArray();
+
+            if(!$array){
+                return $this->response->redirect(site_url('/application'));
+            }
+
+            $data = [
+                'page_title' => 'Applications | TAMU CyberSec Center',
+                'programs' => $array
+            ];
+
+            return view('application/create', $data);
+        }
+        return $this->response->redirect(site_url('/application'));
+
     }
 
     public function new()
     {
+        $user = sessionUser();
+        if(!$user) return $this->response->redirect(site_url('/login'));
+        if($user->hasPermission('admin')) return $this->response->redirect(site_url('/application'));
+
         $method = $this->request->getMethod();
 
         if ($method == "post") {
@@ -66,27 +104,37 @@ class ApplicationController extends BaseController
                 return $this->response->redirect(site_url('program/'));
             }
 
-            // TODO: get UIN from session
-            $this->db->query('INSERT INTO application (program_num, UIN, uncom_cert, com_cert, purpose_statement)
-             VALUES(\''.$formData['program'].'\',230006744,\''.$formData['uncom_cert'].'\',\''.$formData['com_cert'].'\',\''.$formData['purpose_statement'].'\');');
+            $sql =  <<<SQL
+                INSERT INTO application (program_num, UIN, uncom_cert, com_cert, purpose_statement)
+                VALUES('{$formData['program']}',$user->UIN,'{$formData['uncom_cert']}','{$formData['com_cert']}','{$formData['purpose_statement']}');
+            SQL;
+            $this->db->query($sql);
         }
 
-        return $this->response->redirect(site_url('program/'));
+        return $this->response->redirect(site_url('/application'));
 	}
 
     public function edit($app_num, $program_num){
+        $user = sessionUser();
+        if(!$user) return $this->response->redirect(site_url('/login'));
+        if($user->hasPermission('admin')) return $this->response->redirect(site_url('/application'));
+
         $query = $this->db->query('SELECT * FROM application WHERE app_num = \''.$app_num.'\';');
         $programQuery = $this->db->query('SELECT name FROM program WHERE program_num = '.$program_num.';');
         $data = [
             'page_title' => 'Edit Application | TAMU CyberSec Center',
             'app_num' => $app_num,
-            'application' => $query->getResultArray(),
+            'application' => $query->getRowArray(),
             'program' => $programQuery->getRowArray()
         ];
 
         return view('application/edit', $data);
     }
     public function update($app_num){
+        $user = sessionUser();
+        if(!$user) return $this->response->redirect(site_url('/login'));
+        if($user->hasPermission('admin')) return $this->response->redirect(site_url('/application'));
+
         $method = $this->request->getMethod();
         if($method == "post"){
             $formData = $this->request->getPost();
@@ -97,15 +145,11 @@ class ApplicationController extends BaseController
         return $this->response->redirect(site_url('application'));
     }
 
-    // TODO:
-
-    public function documentation(){}
-
-    public function new_doc(){
-
-    }
-
 	public function delete($app_num){
+        $user = sessionUser();
+        if(!$user) return $this->response->redirect(site_url('/login'));
+        if($user->hasPermission('admin')) return $this->response->redirect(site_url('/application'));
+
 		$query = $this->db->query('SELECT * FROM application WHERE app_num = '.$app_num.';');
         $application = $query->getResultArray();
         $programQuery = $this->db->query('SELECT name FROM program WHERE program_num = '.$application[0]['program_num'].';');
@@ -120,11 +164,57 @@ class ApplicationController extends BaseController
 	}
 
 	public function destroy($id){
+        $user = sessionUser();
+        if(!$user) return $this->response->redirect(site_url('/login'));
+        if($user->hasPermission('admin')) return $this->response->redirect(site_url('/application'));
+
         if($id != null){
            $this->db->query('DELETE FROM application WHERE app_num = '.$id.';');
         }
 		
         return $this->response->redirect(site_url('application'));
 	}
+
+    public function review($app_num, $program_num){
+        $user = sessionUser();
+        if(!$user) return $this->response->redirect(site_url('/login'));
+        if($user->hasPermission('student')) return $this->response->redirect(site_url('/application'));
+                
+        $query = $this->db->query('SELECT * FROM application WHERE app_num = \''.$app_num.'\';');
+        $programQuery = $this->db->query('SELECT name FROM program WHERE program_num = '.$program_num.';');
+        $data = [
+            'page_title' => 'Review Application | TAMU CyberSec Center',
+            'app_num' => $app_num,
+            'application' => $query->getRowArray(),
+            'program' => $programQuery->getRowArray()
+        ];
+
+        return view('application/review', $data);
+    }
+
+    public function update_status($app_num){
+        $user = sessionUser();
+        if(!$user) return $this->response->redirect(site_url('/login'));
+        if($user->hasPermission('student')) return $this->response->redirect(site_url('/application'));
+
+        $method = $this->request->getMethod();
+        if($method == "post"){
+
+            $sql;
+            $formData = $this->request->getPost();
+            if($formData['status'] == '0'){
+                $sql = <<<SQL
+                    UPDATE application SET status = 0 WHERE app_num = $app_num;
+                SQL;
+            }elseif($formData['status'] == '1'){
+                $sql = <<<SQL
+                    UPDATE application SET status = 1 WHERE app_num = $app_num;
+                SQL;
+            }
+            $this->db->query($sql);
+
+        }
+        return $this->response->redirect(site_url('application'));
+    }
 
 }
